+++ src/main/java/baseball/domain/strategy/RandomNumberStrategy.java
@@ -1,4 +1,4 @@
-package baseball.domain;
+package baseball.domain.strategy;
 
 @FunctionalInterface
 public interface RandomNumberStrategy {
